+++ CHANGELOG.md
@@ -1,10 +1,19 @@
+## [0.1.4] - 03/03/2023.
+
+* Update API.
+* primaryVariant -> primaryContainer.
+* secondaryVariant -> secondaryContainer.
+
 ## [0.1.3] - 03/03/2021.
+
 * Add secondaryVariant.
 
 ## [0.1.1] - 23/02/2021.
+
 * Update package description and README images to work in pub.dev.
 
 ## [0.1.0] - 23/02/2021.
+
 * Add null-safety.
 
 ## [0.0.1] - 04/10/2020.